# api-service/docs/swagger.yaml
basePath: /
definitions:
  common.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error:
        example: Error message
        type: string
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  request.AlertRuleCreateRequest:
    properties:
      condition_expression:
        type: string
      is_enabled:
        type: boolean
      metric_name:
        type: string
      name:
        maxLength: 64
        type: string
      notification_channels:
        type: string
      rule_type:
        allOf:
        - $ref: '#/definitions/model.AlertRuleType'
        enum:
        - METRIC
        - LOG
        - EVENT
      target_id:
        type: integer
      target_type:
        allOf:
        - $ref: '#/definitions/model.TargetType'
        enum:
        - SERVER
        - APP_INSTANCE
        - WORKFLOW
    required:
    - condition_expression
    - name
    - rule_type
    - target_type
    type: object
...
  response.AlertRecordListResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      records:
        items:
          $ref: '#/definitions/response.AlertRecordResponse'
        type: array
      total:
        type: integer
    type: object
...
  response.AlertRuleListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/response.AlertRuleResponse'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact:
    email: support@websoft9.com
    name: Websoft9 Support
    url: https://www.websoft9.com
  description: Enterprise cloud application management platform API service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Websoft9 API Service
  version: "1.0"
paths:
  /api/v1/alert/records:
    get:
      consumes:
      - application/json
      description: Get a list of alert records with optional filtering
      parameters:
      - description: Page number, default 1
        in: query
        name: page
        type: integer
      - description: Page size, default 20
        in: query
        name: page_size
        type: integer
      - description: Status filter (FIRING, RESOLVED)
        in: query
        name: status
        type: string
      - description: Severity filter
        in: query
        name: severity
        type: string
      - description: Start time filter (ISO8601 format)
        in: query
        name: start_time
        type: string
      - description: End time filter (ISO8601 format)
        in: query
        name: end_time
        type: string
      - description: Alert rule ID filter
        in: query
        name: alert_rule_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.AlertRecordListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Get alert records list
      tags:
      - Alert
...
  /api/v1/alert/rules:
    get:
      consumes:
      - application/json
      description: Get a list of alert rules with optional filtering
      parameters:
      - description: Page number, default 1
        in: query
        name: page
        type: integer
      - description: Page size, default 20
        in: query
        name: page_size
        type: integer
      - description: Rule type filter
        in: query
        name: rule_type
        type: string
      - description: Target type filter
        in: query
        name: target_type
        type: string
      - description: Is enabled filter
        in: query
        name: is_enabled
        type: boolean
      - description: Keyword for name search
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.AlertRuleListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Get alert rules list
      tags:
      - Alert
    post:
      consumes:
      - application/json
      description: Create a new alert rule
      parameters:
      - description: Alert rule create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.AlertRuleCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.AlertRuleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Create alert rule
      tags:
      - Alert
...
  /api/v1/api-tokens/refresh:
    get:
      consumes:
      - application/json
      description: Refresh API token to extend expiration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.APITokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Refresh API token
      tags:
      - API Token Management
  /api/v1/api-tokens/revoke:
    post:
      consumes:
      - application/json
      description: Revoke specified API token
      parameters:
      - description: Revoke API token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RevokeAPITokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Revoke API token
      tags:
      - API Token Management
  /api/v1/audit-logs:
    get:
      description: Get paginated audit log list with multiple filter conditions
      parameters:
      - default: 1
        description: Page Number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page Size
        in: query
        name: page_size
        type: integer
      - description: User ID
        in: query
        name: user_id
        type: integer
      - description: Action Type
        in: query
        name: action
        type: string
      - description: Resource Type
        in: query
        name: resource_type
        type: string
      - description: Resource ID
        in: query
        name: resource_id
        type: integer
      - description: Start Time
        format: date-time
        in: query
        name: start_time
        type: string
      - description: End Time
        format: date-time
        in: query
        name: end_time
        type: string
      - description: IP Address
        in: query
        name: ip_address
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuditLogListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get audit log list
      tags:
      - Audit Log
  /api/v1/audit-logs/{id}:
    get:
      description: Get audit log details by ID
      parameters:
      - description: Audit Log ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuditLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get audit log details
      tags:
      - Audit Log
  /api/v1/audit-logs/export:
    get:
      description: Export audit logs by conditions, supports multiple formats
      parameters:
      - default: csv
        description: Export Format
        enum:
        - csv
        - excel
        - json
        in: query
        name: format
        type: string
      - description: Start Time
        format: date-time
        in: query
        name: start_time
        type: string
      - description: End Time
        format: date-time
        in: query
        name: end_time
        type: string
      - description: User ID
        in: query
        name: user_id
        type: integer
      responses:
        "200":
          description: Export File
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Export audit logs
      tags:
      - Audit Log
  /api/v1/audit-logs/statistics:
    get:
      description: Get audit log statistical analysis data
      parameters:
      - description: Start Time
        format: date-time
        in: query
        name: start_time
        type: string
      - description: End Time
        format: date-time
        in: query
        name: end_time
        type: string
      - default: day
        description: Group By
        enum:
        - hour
        - day
        - week
        - month
        in: query
        name: group_by
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuditLogStatisticsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get audit log statistics
      tags:
      - Audit Log
  /api/v1/auth-config:
    get:
      consumes:
      - application/json
      description: Get current authentication configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.AuthConfigResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get authentication config
      tags:
      - Authentication Config
    put:
      consumes:
      - application/json
      description: Update authentication configuration
      parameters:
      - description: Update auth config request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateAuthConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Update authentication config
      tags:
      - Authentication Config
  /api/v1/auth-config/oauth2-providers:
    get:
      consumes:
      - application/json
      description: Get OAuth2 provider configurations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.OAuth2ProviderResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get OAuth2 providers
      tags:
      - Authentication Config
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email to user
      parameters:
      - description: Forgot password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Forgot password
      tags:
      - Authentication
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, return JWT token
      parameters:
      - description: User login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.UserLoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: User login
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Authentication
  /api/v1/auth/oauth2/login:
    post:
      consumes:
      - application/json
      description: Login with OAuth2 provider
      parameters:
      - description: OAuth2 login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.OAuth2LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.UserLoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: OAuth2 login
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account with email verification
      parameters:
      - description: User registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: User registration
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the 'Bearer ' prefix, e.g. 'Bearer abc123'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"