# api-service/internal/router/router.go
package router

import (
	"api-service/internal/config"
	"api-service/internal/controller"
	serviceInterface "api-service/internal/interface/service"
	"api-service/internal/middleware"
	"api-service/pkg/errors"
	"api-service/pkg/logger"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	ginSwagger "github.com/swaggo/gin-swagger"
	"github.com/swaggo/gin-swagger/swaggerFiles"
)

// Controllers controller collection
type Controllers struct {
	UserController           *controller.UserController
	UserAuthController       *controller.UserAuthController
	I18nController           *controller.I18nController
	RolePermissionController *controller.RolePermissionController
	SecurityController       *controller.SecurityController
	HealthController         *controller.HealthController
	AuditLogController       *controller.AuditLogController
	UserProfileController    *controller.UserProfileController
	SystemConfigController   *controller.SystemConfigController
	AlertController          *controller.AlertController
	TagController            *controller.TagController
	SecretKeyController      *controller.SecretKeyController
	ServerController         *controller.ServerController
	// More controllers can be added
	// AppController  *controller.ApplicationController
}

// SetupRouter sets up router
func SetupRouter(
	controllers *Controllers,
	cfg *config.Config,
	log logger.Logger,
	permissionService serviceInterface.PermissionService,
	apiTokenService serviceInterface.APITokenService,
	auditLogService serviceInterface.AuditLogService,
) *gin.Engine {
	// Set Gin mode
	gin.SetMode(cfg.Server.Mode)

	r := gin.New()

	// Health check
	setupHealthCheck(r, controllers.HealthController)

	// Swagger documentation (without middleware)
	setupSwaggerRoute(r, cfg)

	// Global middleware
	setupMiddleware(r, cfg, log, permissionService, apiTokenService, auditLogService)

	// API routes
	v1 := r.Group("/api/v1")
	setupAPIRoutes(v1, controllers)

	r.NoRoute(func(c *gin.Context) {
		c.JSON(http.StatusNotFound, gin.H{
			"code":    http.StatusNotFound,
			"message": "API endpoint not found",
			"path":    c.Request.URL.Path,
		})
	})

	return r
}

// setupMiddleware sets up global middleware
func setupMiddleware(
	r *gin.Engine,
	cfg *config.Config,
	log logger.Logger,
	permissionService serviceInterface.PermissionService,
	apiTokenService serviceInterface.APITokenService,
	auditLogService serviceInterface.AuditLogService,
) {
	r.Use(middleware.LoggerMiddleware(log))
	r.Use(middleware.CORS())
	r.Use(middleware.I18nMiddleware())
	r.Use(middleware.PermissionMiddleware(permissionService, apiTokenService, cfg, log))
	r.Use(middleware.AuditLogMiddleware(auditLogService, log))
	r.Use(errors.ErrorHandler(log))
}

// setupHealthCheck sets up health check routes
func setupHealthCheck(r *gin.Engine, healthController *controller.HealthController) {
	if healthController == nil {
		// Fallback basic health check if controller is not available
		r.GET("/health", func(c *gin.Context) {
			c.JSON(http.StatusOK, gin.H{
				"status":  "ok",
				"message": "API service running normally",
			})
		})
		return
	}

	// Basic health endpoints
	r.GET("/health", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{
			"status":    "healthy",
			"message":   "API service running normally",
			"timestamp": time.Now().Unix(),
		})
	})

	// Health controller endpoints
	r.GET("/ping", healthController.Ping)
	r.GET("/readiness", healthController.Readiness)
	r.GET("/liveness", healthController.Liveness)

	// Health check API group
	healthGroup := r.Group("/health")
	healthGroup.GET("/system", healthController.SystemHealth)
	healthGroup.GET("/database", healthController.DatabaseHealth)
	healthGroup.GET("/database/stats", healthController.DatabaseStats)
}

// setupSwaggerRoute sets up swagger documentation route
func setupSwaggerRoute(r *gin.Engine, cfg *config.Config) {
	// Only expose swagger in development mode
	if cfg.Server.Mode == "debug" || cfg.Server.Mode == "test" {
		// Add swagger route without middleware
		r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	}
}

// setupAPIRoutes sets up API routes
func setupAPIRoutes(v1 *gin.RouterGroup, controllers *Controllers) {
	// Routes requiring JWT authentication
	protected := v1.Group("/")
	setupUserAuthRoutes(v1, controllers.UserAuthController)
	setupI18nRoutes(v1, protected, controllers.I18nController)
	setupUserRoutes(protected, controllers.UserController)
	setupRoleRoutes(protected, controllers.RolePermissionController)
	setupPermissionRoutes(protected, controllers.RolePermissionController)
	setupAPITokenRoutes(protected, controllers.SecurityController)
	setupTwoFactorRoutes(protected, controllers.SecurityController)
	setupAuthConfigRoutes(protected, controllers.SecurityController)
	setupAuditLogRoutes(protected, controllers.AuditLogController)
	setupUserProfileRoutes(protected, controllers.UserProfileController)
	setupSystemConfigRoutes(protected, controllers.SystemConfigController)
	setupTagRoutes(protected, controllers.TagController)
	setupAlertRoutes(protected, controllers.AlertController)
	setupSecretKeyRoutes(protected, controllers.SecretKeyController)
	setupServerRoutes(protected, controllers.ServerController)
}

// setupUserRoutes sets up user related routes
func setupUserAuthRoutes(v1 *gin.RouterGroup, userAuthController *controller.UserAuthController) {
	// Authentication related routes (no JWT verification required)
	auth := v1.Group("/auth")
	auth.POST("/register", userAuthController.Register)
	auth.POST("/login", userAuthController.Login)
	auth.POST("/logout", userAuthController.Logout)
	auth.POST("/forgot-password", userAuthController.ForgotPassword)
	auth.GET("/reset-password", userAuthController.ShowResetPasswordForm)
	auth.POST("/reset-password", userAuthController.ResetPassword)
	auth.GET("/verify-email", userAuthController.VerifyEmail)
	auth.POST("/resend-verification", userAuthController.ResendVerificationEmail)
	auth.POST("/oauth2/login", userAuthController.OAuth2Login)
}

// setupI18nRoutes sets up internationalization routes
func setupI18nRoutes(v1, protected *gin.RouterGroup, i18nController *controller.I18nController) {
	if i18nController == nil {
		return
	}

	// i18n related routes (no JWT verification required)
	i18nGroup := v1.Group("/i18n")
	i18nGroup.GET("/languages", i18nController.GetLanguages)

	// Protected i18n routes (JWT verification required)
	protectedI18n := protected.Group("/i18n")
	protectedI18n.POST("/switch-language", i18nController.SwitchLanguage)
}

// setupUserRoutes sets up user management routes
func setupUserRoutes(protected *gin.RouterGroup, userController *controller.UserController) {
	users := protected.Group("/users")

	// User management operations (admin permissions required)
	users.POST("", userController.CreateUser)
	users.GET("", userController.ListUsers)
	users.GET("/:id", userController.GetUser)
	users.PUT("/:id", userController.UpdateUser)
	users.PUT("/:id/status", userController.UpdateUserStatus)
	users.PUT("/:id/password", userController.UpdateUserPassword)
	users.DELETE("/:id", userController.DeleteUser)
}

// setupRoleRoutes sets up role management routes
func setupRoleRoutes(protected *gin.RouterGroup, roleController *controller.RolePermissionController) {
	if roleController == nil {
		return
	}

	roles := protected.Group("/roles")
	roles.GET("", roleController.ListRoles)
	roles.POST("", roleController.CreateRole)
	roles.GET("/:id", roleController.GetRole)
	roles.PUT("/:id", roleController.UpdateRole)
	roles.DELETE("/:id", roleController.DeleteRole)

	// Role permission management
	roles.POST("/:id/permissions", roleController.AssignPermissions)
	roles.DELETE("/:id/permissions", roleController.RemovePermissions)
	roles.GET("/:id/users", roleController.GetRoleUsers)
}

// setupPermissionRoutes sets up permission management routes
func setupPermissionRoutes(protected *gin.RouterGroup, permissionController *controller.RolePermissionController) {
	if permissionController == nil {
		return
	}

	permissions := protected.Group("/permissions")
	permissions.GET("", permissionController.ListPermissions)
	permissions.GET("/tree", permissionController.GetPermissionTree)
	permissions.POST("", permissionController.CreatePermission)
	permissions.GET("/:id", permissionController.GetPermission)
	permissions.PUT("/:id", permissionController.UpdatePermission)
	permissions.DELETE("/:id", permissionController.DeletePermission)
	permissions.GET("/:id/roles", permissionController.GetPermissionRoles)
}

// setupAPITokenRoutes sets up API token management routes
func setupAPITokenRoutes(protected *gin.RouterGroup, securityController *controller.SecurityController) {
	if securityController == nil {
		return
	}

	apiTokens := protected.Group("/api-tokens")
	apiTokens.POST("/revoke", securityController.RevokeAPIToken)
	apiTokens.GET("/refresh", securityController.RefreshAPIToken)
}

// setupTwoFactorRoutes sets up two-factor authentication routes
func setupTwoFactorRoutes(protected *gin.RouterGroup, securityController *controller.SecurityController) {
	if securityController == nil {
		return
	}

	// Two-factor authentication routes under users
	twoFactor := protected.Group("/two-factor")
	twoFactor.GET("", securityController.GetTwoFactorStatus)
	twoFactor.POST("/enable", securityController.EnableTOTP)
	twoFactor.POST("/confirm", securityController.ConfirmTOTP)
	twoFactor.POST("/disable", securityController.DisableTwoFactor)
	twoFactor.POST("/verify", securityController.VerifyTwoFactor)
	twoFactor.POST("/totp/generate", securityController.GenerateTOTPSecret)
}

// setupAuthConfigRoutes sets up authentication config routes
func setupAuthConfigRoutes(protected *gin.RouterGroup, securityController *controller.SecurityController) {
	if securityController == nil {
		return
	}

	// Authentication config routes
	authConfig := protected.Group("/auth-config")
	authConfig.GET("", securityController.GetAuthConfig)
	authConfig.PUT("", securityController.UpdateAuthConfig)
	authConfig.GET("/oauth2-providers", securityController.GetOAuth2Providers)
}

// setupAuditLogRoutes sets up audit log routes
func setupAuditLogRoutes(protected *gin.RouterGroup, auditLogController *controller.AuditLogController) {
	if auditLogController == nil {
		return
	}

	// Audit log routes
	auditLogs := protected.Group("/audit-logs")
	auditLogs.GET("", auditLogController.ListAuditLogs)
	auditLogs.GET("/:id", auditLogController.GetAuditLog)
	auditLogs.GET("/statistics", auditLogController.GetAuditLogStatistics)
	auditLogs.GET("/export", auditLogController.ExportAuditLogs)
}

// setupUserProfileRoutes sets up user profile routes
func setupUserProfileRoutes(protected *gin.RouterGroup, userProfileController *controller.UserProfileController) {
	if userProfileController == nil {
		return
	}

	profile := protected.Group("/profile")
	profile.GET("", userProfileController.GetProfile)
	profile.PUT("", userProfileController.UpdateProfile)
	profile.PUT("/password", userProfileController.ChangePassword)
	profile.GET("/login-history", userProfileController.GetLoginHistories)
	profile.GET("/notification-settings", userProfileController.GetNotificationSettings)
	profile.PUT("/notification-settings", userProfileController.UpdateNotificationSettings)
	profile.GET("/security-settings", userProfileController.GetSecuritySettings)
	profile.PUT("/security-settings", userProfileController.UpdateSecuritySettings)
}

// setupSystemConfigRoutes sets up system configuration routes
func setupSystemConfigRoutes(protected *gin.RouterGroup, systemConfigController *controller.SystemConfigController) {
	if systemConfigController == nil {
		return
	}

	// System configuration routes
	systemConfigs := protected.Group("/system-configs")
	systemConfigs.GET("", systemConfigController.ListSystemConfigs)
	systemConfigs.PUT("", systemConfigController.BatchUpdateSystemConfigs)
	systemConfigs.GET("/basic", systemConfigController.ListBasicConfigs)
	systemConfigs.GET("/security", systemConfigController.ListSecurityConfigs)
	systemConfigs.GET("/email", systemConfigController.ListEmailConfigs)
	systemConfigs.POST("/smtp/test", systemConfigController.TestSMTP)
}

// setupTagRoutes sets up tag management routes
func setupTagRoutes(protected *gin.RouterGroup, tagController *controller.TagController) {
	if tagController == nil {
		return
	}

	// Tag basic management endpoints
	tags := protected.Group("/tags")
	tags.GET("", tagController.ListTags)
	tags.POST("", tagController.CreateTag)
	tags.GET("/search", tagController.SearchTags)
	tags.GET("/:id", tagController.GetTag)
	tags.PUT("/:id", tagController.UpdateTag)
	tags.DELETE("/:id", tagController.DeleteTag)

	// Tag-resource association endpoints
	tags.POST("/assign", tagController.AssignTags)
	tags.POST("/replace", tagController.ReplaceTags)
	tags.POST("/unassign", tagController.UnassignTags)
	tags.GET("/taggings", tagController.GetResourceTags)
	tags.GET("/taggings/search", tagController.SearchResourcesByTags)
}

// setupAlertRoutes registers all alert related routes
func setupAlertRoutes(protected *gin.RouterGroup, alertController *controller.AlertController) {
	alertGroup := protected.Group("/alert")

	// Alert rules routes
	rulesGroup := alertGroup.Group("/rules")
	rulesGroup.GET("", alertController.GetAlertRules)          // Get list of alert rules
	rulesGroup.POST("", alertController.CreateAlertRule)       // Create a new alert rule
	rulesGroup.GET("/:id", alertController.GetAlertRule)       // Get a single alert rule by ID
	rulesGroup.PUT("/:id", alertController.UpdateAlertRule)    // Update an alert rule
	rulesGroup.DELETE("/:id", alertController.DeleteAlertRule) // Delete an alert rule

	recordsGroup := alertGroup.Group("/records")
	recordsGroup.GET("", alertController.GetAlertRecords)                        // Get list of alert records
	recordsGroup.PUT("/:id/acknowledge", alertController.AcknowledgeAlertRecord) // Acknowledge an alert
	recordsGroup.PUT("/:id/resolve", alertController.ResolveAlertRecord)         // Resolve an alert
}

// setupSecretKeyRoutes sets up secret key management routes
func setupSecretKeyRoutes(protected *gin.RouterGroup, secretKeyController *controller.SecretKeyController) {
	if secretKeyController == nil {
		return
	}

	// Secret key routes
	secretKeys := protected.Group("/secret-keys")
	secretKeys.GET("", secretKeyController.ListSecretKeys)              // GET /api/v1/secret-keys
	secretKeys.POST("", secretKeyController.CreateSecretKey)            // POST /api/v1/secret-keys
	secretKeys.GET("/export", secretKeyController.ExportSecretKeys)     // GET /api/v1/secret-keys/export
	secretKeys.GET("/:id", secretKeyController.GetSecretKey)            // GET /api/v1/secret-keys/{id}
	secretKeys.GET("/:id/value", secretKeyController.GetSecretKeyValue) // GET /api/v1/secret-keys/{id}/value
	secretKeys.PUT("/:id", secretKeyController.UpdateSecretKey)         // PUT /api/v1/secret-keys/{id}
	secretKeys.DELETE("/:id", secretKeyController.DeleteSecretKey)      // DELETE /api/v1/secret-keys/{id}
}

// setupServerRoutes sets up server management routes
func setupServerRoutes(protected *gin.RouterGroup, serverController *controller.ServerController) {
	if serverController == nil {
		return
	}

	// Server basic management endpoints
	servers := protected.Group("/servers")
	servers.GET("", serverController.ListServers)                // GET /api/v1/servers
	servers.POST("", serverController.CreateServer)              // POST /api/v1/servers
	servers.GET("/:id", serverController.GetServer)              // GET /api/v1/servers/{id}
	servers.PUT("/:id", serverController.UpdateServer)           // PUT /api/v1/servers/{id}
	servers.DELETE("/:id", serverController.DeleteServer)        // DELETE /api/v1/servers/{id}
	servers.GET("/:id/status", serverController.GetServerStatus) // GET /api/v1/servers/{id}/status (6.3.6)

	// Server batch operations
	servers.POST("/status", serverController.CheckServersStatus)    // POST /api/v1/servers/status (6)
	servers.POST("/actions", serverController.ExecuteServerActions) // POST /api/v1/servers/actions (7)

	// Server file management endpoints
	servers.POST("/:id/files", serverController.UploadFile)           // POST /api/v1/servers/{id}/files (8)
	servers.GET("/:id/files/download", serverController.DownloadFile) // GET /api/v1/servers/{id}/files/download (9)
	servers.DELETE("/:id/files", serverController.DeleteFile)         // DELETE /api/v1/servers/{id}/files (10)
}